adjusted the initialization
Although this script is not really needed,
as we had seen yesterday,
this is a working version.

(The idea to use Julia was apparently not good:
Currently there are problems with its Tar.jl package
for unpacking some of the archives.
Also the functionality for uncompressing different formats
looks not user friendly,
and support for .zip archives seems to be missing.)